pneumonia-xray-github/requirements.txt

torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.25.0
streamlit>=1.28.0
flake8>=6.1.0
pytest>=7.4.0
